// src/utils/Models/Trailer.tsx
export type TrailerDetails = {
    iso_639_1: string;
    iso_3166_1: string;
    name: string;
    key: string;
    site: string;
    size: number;
    type: string;
    official: boolean;
    published_at: Date;
    id: string;
  }
  
  export type Trailer = {
    id: number;
    results: TrailerDetails[];
  }
  